Remove smart pointers from MeshCatalog
